# improve bucket.Get requests performance